# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Xanda Cross unified inbox API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /messages/send:
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message on any connected platform
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageSendInput"
      responses:
        "200":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageSendResult"
  /xan/chat:
    post:
      operationId: chatWithXan
      tags: [xan]
      summary: Send a message to Xan AI assistant
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/XanChatInput"
      responses:
        "200":
          description: Xan response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/XanChatResponse"
  /xan/draft/{conversationId}:
    post:
      operationId: regenerateDraft
      tags: [xan]
      summary: Regenerate Xan AI draft reply for a conversation
      security: [{ bearerAuth: [] }]
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: New draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DraftReply"
  /xan/history:
    get:
      operationId: getXanHistory
      tags: [xan]
      summary: Get Xan chat history
      security: [{ bearerAuth: [] }]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Chat history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/XanHistoryList"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    adminAuth:
      type: http
      scheme: bearer
  schemas:
...
    ConversationSummary:
      type: object
      properties:
        id:
          type: string
        platform:
          type: string
        contactName:
          type: string
        contactAvatarUrl:
          type: ["string", "null"]
        contactId:
          type: ["string", "null"]
        topicLabel:
          type: ["string", "null"]
        headline:
          type: ["string", "null"]
        preview:
          type: ["string", "null"]
        priority:
          type: string
        isRead:
          type: boolean
        lastMessageAt:
          type: string
        unreadCount:
          type: integer
      required: [id, platform, contactName, priority, isRead, lastMessageAt, unreadCount]
...
    DraftReply:
      type: object
      properties:
        draft:
          type: string
        conversationId:
          type: string
      required: [draft, conversationId]
    MessageSendInput:
      type: object
      properties:
        conversationId:
          type: string
        platform:
          type: string
        body:
          type: string
        accountId:
          type: ["string", "null"]
      required: [conversationId, platform, body]
    MessageSendResult:
      type: object
      properties:
        success:
          type: boolean
        messageId:
          type: ["string", "null"]
      required: [success]
    XanChatInput:
      type: object
      properties:
        message:
          type: string
        conversationId:
          type: ["string", "null"]
      required: [message]
    XanChatResponse:
      type: object
      properties:
        response:
          type: string
        type:
          type: string
        relatedConversations:
          type: array
          items:
            $ref: "#/components/schemas/ConversationSummary"
        draft:
          type: ["string", "null"]
      required: [response, type]
    XanMessage:
      type: object
      properties:
        id:
          type: string
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
      required: [id, role, content, createdAt]
    XanHistoryList:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/XanMessage"
      required: [messages]
...