# lib/api-zod/src/generated/types/xanChatInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Xanda Cross unified inbox API
 * OpenAPI spec version: 0.1.0
 */

export interface XanChatInput {
  message: string;
  /** @nullable */
  conversationId?: string | null;
}