# lib/api-zod/src/generated/types/xanHistoryList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Xanda Cross unified inbox API
 * OpenAPI spec version: 0.1.0
 */
import type { XanMessage } from './xanMessage';

export interface XanHistoryList {
  messages: XanMessage[];
}